refactor: move otel configuration to observability package
