+++ server/src/main/kotlin/fi/oph/kitu/observability/OpenTelemetryTracerConfig.kt
@@ -1,4 +1,4 @@
-package fi.oph.kitu.logging
+package fi.oph.kitu.observability
 
 import io.opentelemetry.api.OpenTelemetry
 import io.opentelemetry.api.trace.Tracer
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-class OpenTelemetryConfig {
+class OpenTelemetryTracerConfig {
     @Bean
     fun tracer(openTelemetry: OpenTelemetry): Tracer = openTelemetry.getTracer("Kitu")
 }
