+++ server/src/main/kotlin/fi/oph/kitu/yki/suoritukset/error/YkiSuoritusErrorService.kt
@@ -6,8 +6,8 @@ import fi.oph.kitu.csvparsing.setSerializationErrorToAttributes
 import fi.oph.kitu.findAllSorted
 import fi.oph.kitu.jdbc.replaceAll
 import fi.oph.kitu.logging.AuditLogger
-import fi.oph.kitu.logging.setAttribute
-import fi.oph.kitu.logging.use
+import fi.oph.kitu.observability.setAttribute
+import fi.oph.kitu.observability.use
 import fi.oph.kitu.yki.suoritukset.YkiSuoritusCsv
 import io.opentelemetry.api.trace.Tracer
 import io.opentelemetry.instrumentation.annotations.WithSpan
